Fix from Tom Shanley.  Tom's comments:
I noticed that even with the example sentence in the test program (The quick brown fox...) some words would not get tagged.  This looked wrong as I understood that Brill's algorithm would tag every word.  I tracked this problem down to the line
 
if ((SubRule.Length == 5))
 
in BrillTagger.DoLexicalTagging().  As I read it, rules with a meaningful value in SubRule[4] are always of length 6 when split (I guess the 'x' is just a placeholder to simplify Brill's C code?). As it was, a lexical rule such as WSJ line 63 ("JJ n fdeletesuf 1 VBN x") would always assign a null to the tag when it fired. When I changed that code to ">= 5", I no longer see the untagged words, and the tags make sense, subject to oddities in the different rule sets and lexicons; that WSJ rule in particular decides "brown" is the word "brow" with suffix "n", with the result that the final tag is actually wrong. (Actually I think "== 6" would be more correct, but >=5 seems to work OK.)